[refactor] simplify pygame viewer
