+++ examples/colorizer.py
@@ -106,10 +106,9 @@ def setup_gan(config, inputs, args):
 
     tf.train.start_queue_runners(sess=gan.session)
 
-    GlobalViewer.enable()
     config_name = args.config
     title = "[hypergan] colorizer " + config_name
-    GlobalViewer.window.set_title(title)
+    #GlobalViewer.window.set_title(title)
 
     return gan
