+++ examples/static.py
@@ -55,7 +55,6 @@ def setup_gan(config, inputs, args):
 
     tf.train.start_queue_runners(sess=gan.session)
 
-    GlobalViewer.enable()
     config_name = args.config
     title = "[hypergan] static " + config_name
     GlobalViewer.window.set_title(title)
