+++ hypergan/pygame_viewer.py
@@ -3,7 +3,6 @@ Opens a window that displays an image.
 Usage:
 
     from viewer import GlobalViewer
-    GlobalViewer.enable()
     GlobalViewer.update(image)
 
 Delays loading Gtk and friends until enable() is called.
@@ -11,21 +10,18 @@ Delays loading Gtk and friends until enable() is called.
 import numpy as np
 import pygame
 
-class SdlViewer:
-    def __init__(self, size = (2048, 512)):
-        self.size = size
+class PygameViewer:
 
-    def enable(self):
-        self.screen = pygame.display.set_mode(self.size)
+    def __init__(self):
+        self.screen = None
 
     def update(self, image):
-        print("preImage is ",image.shape)
         image = np.transpose(image, [1, 0,2])
-        print("Image is ",image.shape)
-        surface = pygame.Surface(self.size)
+        size = [image.shape[0], image.shape[1]]
+        if not self.screen:
+            self.screen = pygame.display.set_mode(size)
+        surface = pygame.Surface(size)
         pygame.surfarray.blit_array(surface, image)
-        #surface = pygame.transform.rotate(surface, 90)
         self.screen.blit(surface, (0,0))
         pygame.display.flip()
 
-GlobalViewer = SdlViewer()
