+++ hypergan/samplers/base_sampler.py
@@ -1,6 +1,6 @@
 import numpy as np
 from PIL import Image
-from hypergan.sdl_viewer import GlobalViewer
+from hypergan.viewer import GlobalViewer
 
 class BaseSampler:
     def __init__(self, gan, samples_per_row=8, session=None):
