+++ hypergan/samplers/random_walk_sampler.py
@@ -14,22 +14,27 @@ class RandomWalkSampler(BaseSampler):
 
     def _sample(self):
         gan = self.gan
-        print("UNIFORM")
         z_t = gan.uniform_encoder.sample
         inputs_t = gan.inputs.x
 
         if self.z is None:
-            self.z = gan.uniform_encoder.sample.eval()
-            self.target = gan.uniform_encoder.sample.eval()
+            self.z = gan.uniform_encoder.sample.eval()/2
+            direct = gan.uniform_encoder.sample.eval()[0]/2
+            direct = np.reshape(direct, [1, direct.shape[0]])
+            self.direction = np.tile(direct, [self.z.shape[0], 1])
             self.input = gan.session.run(gan.inputs.x)
 
         if self.step > self.steps:
-            self.z = self.target
-            self.target = gan.uniform_encoder.sample.eval()
+            self.z = np.minimum(self.z+self.direction, 1)
+            self.z = np.maximum(self.z, -1)
+            self.direction = gan.uniform_encoder.sample.eval()
+
             self.step = 0
 
         percent = float(self.step)/self.steps
-        z_interp = self.z*(1.0-percent) + self.target*percent
+        z_interp = self.z + self.direction*percent
+        z_interp = np.minimum(z_interp, 1)
+        z_interp = np.maximum(z_interp, -1)
         self.step+=1
 
         g=tf.get_default_graph()
